Rename to more consistent casing
